extract filter into a separate function, use in both components
